requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.0
# TIGRE: pip install may fail - use official install first: https://github.com/CERN/TIGRE/blob/master/Frontispiece/python_installation.md
tigre>=2.0.0
scipy>=1.7.0
matplotlib>=3.3.0
numpy-stl>=2.16.0
scikit-image>=0.19.0
numpy2ometiff>=0.1.0
